// extension/src/shared/types/react-types.ts
export enum WorkTag {
	FunctionComponent = 0,
	ClassComponent = 1,
	IndeterminateComponent = 2, // ? Before we know whether it is function or class
	HostRoot = 3, // ? Root of a host tree. Could be nested inside another node.
	HostPortal = 4, // ? A subtree. Could be an entry point to a different renderer.
	HostComponent = 5,
	HostText = 6,
	Fragment = 7,
	Mode = 8,
	ContextConsumer = 9,
	ContextProvider = 10,
	ForwardRef = 11,
	Profiler = 12,
	SuspenseComponent = 13,
	MemoComponent = 14,
	SimpleMemoComponent = 15,
	LazyComponent = 16,
	IncompleteClassComponent = 17,
	DehydratedFragment = 18,
	SuspenseListComponent = 19,
	ScopeComponent = 21,
	OffscreenComponent = 22,
	LegacyHiddenComponent = 23,
	CacheComponent = 24,
	TracingMarkerComponent = 25,
	HostHoistable = 26,
	HostSingleton = 27,
}